% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pNMF_func.R
\name{PNMFfun}
\alias{PNMFfun}
\title{Fast Projective Nonnegative Matrix Factorization Realizatiton based on euclidean distance / KL Divergence / Discriminant pNMF.}
\usage{
PNMFfun(
  X,
  rank = 10,
  tol = 1e-05,
  maxIter = 5000,
  verboseN = FALSE,
  zerotol = 1e-10,
  method = "KL",
  label = NULL,
  mu = 1,
  lambda = 0.01,
  seed = 123
)
}
\arguments{
\item{X}{Input data matrix, where rows represent genes, columns represent cells.}

\item{rank}{Specification of the factorization rank.}

\item{tol}{A threshold below which would be considered converged.}

\item{maxIter}{Number of max iteration times.}

\item{verboseN}{A boolean value indicating whether to print number of iterations.}

\item{zerotol}{A threshold on basis loadings below which would be considered zero.}

\item{method}{A character string indicating which method to be used. One of "EucDist", "KL", or "DPNMF".}

\item{label}{A character vector indicating the cluster type for each cell. Required only when method = "DPNMF".}

\item{mu}{A numerical value.}

\item{lambda}{A numerical value.}

\item{seed}{random seed}
}
\value{
A list of fitted NMF model, our pNMF basis, and mapped scores.
}
\description{
Fast Projective Nonnegative Matrix Factorization Realizatiton based on euclidean distance / KL Divergence / Discriminant pNMF.
}
\examples{
k <- matrix(rpois(600, lambda=5), nrow=20, ncol=30)
estim <- PNMFfun(k, rank=3, method="EucDist")



}
