\name{QuickPNMFs-package}
\alias{QuickPNMFs-package}
\alias{QuickPNMFs}
\docType{package}
\title{\packageTitle{QuickPNMFs}}
\description{\packageDescription{QuickPNMFs}}
\section{Package Content}{\packageIndices{QuickPNMFs}}
\author{\packageAuthor{QuickPNMFs}}
\section{Maintainer}{\packageMaintainer{QuickPNMFs}}
\keyword{package}
